move NoteFields to NoteForm children components folder
