+++ lib/Notes/NoteForm/NoteForm.js
@@ -18,9 +18,8 @@ import {
   Button,
 } from '@folio/stripes-components';
 
+import NoteFields from './components/NoteFields';
 import NavigationModal from '../NavigationModal';
-
-import NoteFields from '../NoteFields';
 import AssignmentsList from '../AssignmentsList';
 
 import styles from './NoteForm.css';
