+++ lib/Notes/NoteForm/components/NoteFields/NoteFields.js
@@ -11,7 +11,7 @@ import {
   Editor,
 } from '@folio/stripes-components';
 
-import { noteTypesShape } from '../NoteForm/noteShapes';
+import { noteTypesShape } from '../../noteShapes';
 
 export default class NoteFields extends Component {
   static propTypes = {
